+++ tests/test_version.py
@@ -1,0 +1,5 @@
+from release_exporter import version
+
+
+def test_vesion():
+    assert type(version.__version__) == str